# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Bank API
  version: v1
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - BearerToken: []
...
paths:
  "/api/users/{id}":
    get:
      tags:
        - User Management API
      summary: Получение пользователя по ID
      operationId: getById
      parameters:
        - name: id
          in: path
          description: ID пользователя
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Найденный пользователь
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '404':
          description: Пользователь не найден
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
    put:
      tags:
        - User Management API
      summary: Обновление данных пользователя
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: ID пользователя
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserUpdateRequest"
        required: true
      responses:
        '200':
          description: Пользователь обновлён
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '404':
          description: Пользователь не найден
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
    delete:
      tags:
        - User Management API
      summary: Удаление пользователя по ID
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: ID пользователя
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Пользователь удалён
        '404':
          description: Пользователь не найден
...
  "/api/users":
    get:
      tags:
        - User Management API
      summary: Получение списка всех пользователей
      operationId: getAll
      responses:
        '200':
          description: Список пользователей
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
    post:
      tags:
        - User Management API
      summary: Создание нового пользователя
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserCreateRequest"
        required: true
      responses:
        '201':
          description: Пользователь создан
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '400':
          description: Ошибка валидации запроса
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
  "/api/cards/user/{id}/block-request":
    post:
      tags:
        - Card User API
      summary: Отправка запроса на блокировку карты
      operationId: requestBlock
      parameters:
        - name: id
          in: path
          description: ID карты
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Запрос успешно отправлен
        '403':
          description: Доступ запрещён
        '404':
          description: Карта не найдена
  "/api/cards/user/transfer":
    post:
      tags:
        - Card User API
      summary: Перевод средств между картами
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TransferRequest"
        required: true
      responses:
        '200':
          description: Перевод выполнен
        '400':
          description: Неверные данные
        '403':
          description: Доступ запрещён
...
  "/api/cards/user":
    get:
      tags:
        - Card User API
      summary: Получение списка своих карт
      operationId: getMyCards
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: pageable
          in: query
          required: true
          schema:
            "$ref": "#/components/schemas/Pageable"
      responses:
        '200':
          description: Список карт пользователя
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardResponse"
  "/api/cards/user/{id}":
    get:
      tags:
        - Card User API
      summary: Получение информации о карте по ID
      operationId: getCardById
      parameters:
        - name: id
          in: path
          description: ID карты
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о карте
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardResponse"
        '403':
          description: Доступ запрещён
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardResponse"
        '404':
          description: Карта не найдена
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardResponse"
  "/api/cards/user/{id}/balance":
    get:
      tags:
        - Card User API
      summary: Получение баланса по карте
      operationId: getBalance
      parameters:
        - name: id
          in: path
          description: ID карты
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Баланс карты
          content:
            "*/*":
              schema:
                type: number
        '403':
          description: Доступ запрещён
          content:
            "*/*":
              schema:
                type: number
        '404':
          description: Карта не найдена
          content:
            "*/*":
              schema:
                type: number
  "/api/cards/user/transfer/all":
    get:
      tags:
        - Card User API
      summary: История переводов пользователя
      operationId: getAllTransfer
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            "$ref": "#/components/schemas/Pageable"
      responses:
        '200':
          description: Список переводов
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TransferHistoryResponse"
...
components:
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        role:
          type: string
          enum:
            - ROLE_ADMIN
            - ROLE_USER
        createdAt:
          type: string
          format: date-time
    UserUpdateRequest:
      type: object
      properties:
        fullName:
          type: string
        phoneNumber:
          type: string
        password:
          type: string
...
    UserCreateRequest:
      type: object
      properties:
        fullName:
          type: string
          minLength: 1
        email:
          type: string
        phoneNumber:
          type: string
        password:
          type: string
          minLength: 1
        role:
          type: string
          enum:
            - ROLE_ADMIN
            - ROLE_USER
      required:
        - fullName
        - password
        - role
    TransferRequest:
      type: object
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number
          minimum: 0.01
      required:
        - amount
        - fromCardId
        - toCardId
...
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        expirationDate:
          type: string
          format: date
        status:
          type: string
        balance:
          type: number
        card_number:
          type: string
        owner:
          type: string
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    TransferHistoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        senderCardId:
          type: integer
          format: int64
        receiverCardId:
          type: integer
          format: int64
        amount:
          type: number
        transferredAt:
          type: string
          format: date-time
...
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT